fix: Separated Modal from Table because table has no responsibility to manage Modal
