+++ frontend/context/BillContext.jsx
@@ -24,6 +24,7 @@ export function BillProvider({ children }) {
   }
 
   const createItemKey = () => {
+    // Generating Unique Keys
     let key = Number(Date.now());
     return key
   }
